test: #31 docList 레이아웃 테스트
