+++ frontend/src/components/docList/index.ts
@@ -1,0 +1,1 @@
+export * from './DocList';
