+++ frontend/src/pages/MainPage.tsx
@@ -1,38 +1,40 @@
 import React, { useState, useEffect } from 'react';
+import { DocList } from '../components/docList';
 import { DocListItem } from '../components/docListItem';
 
-interface docListItemState {
-  id: string
-  title: string
-  lastVisited: string
-  role: string,
-}
+// interface docListItemState {
+//   id: string
+//   title: string
+//   lastVisited: string
+//   role: string,
+// }
 
 const MainPage = () => {
-  const [docList, setDocList] = useState<docListItemState[]>([]);
+  // const [docList, setDocList] = useState<docListItemState[]>([]);
 
-  useEffect(() => {
-    const fetchDocList = async () => {
-      // TODO : request docList from API server
-      return new Promise((resolve) => {
-        resolve(setDocList(() => [{id:'1234',
-          title: '알고리즘 스터디 기록',
-          lastVisited: '2022-11-10',
-          role: 'onwer',  
-        }, {id:'5678',
-          title: 'Untitled',
-          lastVisited: '2022-11-19',
-          role: 'edit'}]));
-      });
-    };
-    fetchDocList();
-  }, []);
+  // useEffect(() => {
+  //   const fetchDocList = async () => {
+  //     // TODO : request docList from API server
+  //     return new Promise((resolve) => {
+  //       resolve(setDocList(() => [{id:'1234',
+  //         title: '알고리즘 스터디 기록',
+  //         lastVisited: '2022-11-10',
+  //         role: 'onwer',  
+  //       }, {id:'5678',
+  //         title: 'Untitled',
+  //         lastVisited: '2022-11-19',
+  //         role: 'edit'}]));
+  //     });
+  //   };
+  //   fetchDocList();
+  // }, []);
 
   return (
     <div>
-      {docList.length ? docList.map(doc => {
+      {/* {docList.length ? docList.map(doc => {
         return <DocListItem key={doc.id} id={doc.id} title={doc.title} lastVisited={doc.lastVisited} role={doc.role} />;
-      }) : <div>...loading</div>}
+      }) : <div>...loading</div>} */}
+      <DocList />
     </div>
   );
 };
